backend/requirements.txt
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
coverage==7.7.1
cycler==0.12.1
Cython==3.0.12
easydict==1.13
faiss-cpu==1.10.0
fastapi==0.115.12
flatbuffers==25.2.10
fonttools==4.56.0
greenlet==3.1.1
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
insightface==0.7.3
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
onnx==1.17.0
onnxruntime==1.21.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
prettytable==3.16.0
protobuf==6.30.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.0.2
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
stringzilla==3.12.3
sympy==1.13.3
threadpoolctl==3.6.0
tifffile==2025.3.13
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wheel==0.45.1
psutil>=5.9.0
